requirements.txt
fastapi==0.115.6
sqlmodel==0.0.22
openai==1.57.4
psycopg2-binary==2.9.10
uvicorn==0.34.0
python-multipart==0.0.20
pydantic==2.10.3
pydantic-settings==2.7.1
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sse-starlette==1.6.5
requests==2.32.3
python-dotenv==1.0.1